requirements.txt
pandas>=2.0.0
numpy>=1.24.0
google-genai>=0.2.0
scikit-learn>=1.3.0
krippendorff>=0.6.1
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

